MEDIUM: automatically fallback to host list when not specified
When a scenario uses "." as its host name (including generic ones
built on the fly), then the list of hosts declared by "-h" will be
used and rouhgly evenly spread between clients and processes. This
allows to aim the load at a list of multiple host:ports instead of
a single one.