MEDIUM: adjust rlim_nofile for all users, not just root
Instead of checking for uid==0 prior to performing the change, let's do
it all the time. Most OSes nowadays have a hard limit that's well over
the current one and this will make sure we can always use what we need.
The limit is only set if it's too low, it's never reduced.